Improve test coverage for core/property to meet CI thresholds
- Added tests for recursion limits in serialization (serialize, serialized_size, estimated_heap_size) to cover error paths in `serialization.rs`.
- Added tests for `PropertyMapBuilder` methods (`insert_by_key`, `remove_by_key`, `try_insert`, `try_remove`) to cover `map.rs` functions.
- This addresses the `fail-under-functions` coverage drop detected in CI (87.88% < 88%).
