--- src/core/property/map.rs
@@ -1,446 +1,0 @@
-use std::collections::HashMap;
-use std::sync::Arc;
-
-use super::types::*;
-use crate::core::interning::GLOBAL_INTERNER;
-use crate::utils::error::{Result, StorageError};
-
-/// A map of property keys to values with copy-on-write semantics.
-///
-/// The underlying HashMap is wrapped in an Arc, making clones very cheap
-/// (just incrementing a reference count). This enables efficient sharing
-/// of unchanged properties across versions.
-#[derive(Debug, Clone, PartialEq)]
-pub struct PropertyMap {
-    pub(crate) inner: Arc<HashMap<PropertyKey, PropertyValue>>,
-    /// Cached serialized size in bytes.
-    pub(crate) cached_size: usize,
-}
-
-impl PropertyMap {
-    /// Create a new empty property map.
-    pub fn new() -> Self {
-        PropertyMap {
-            inner: Arc::new(HashMap::new()),
-            cached_size: 4, // 4 bytes for the count field (0)
-        }
-    }
-
-    /// Create a property map with the specified capacity.
-    pub fn with_capacity(capacity: usize) -> Self {
-        PropertyMap {
-            inner: Arc::new(HashMap::with_capacity(capacity)),
-            cached_size: 4, // 4 bytes for the count field (0)
-        }
-    }
-
-    /// Get a property value by key.
-    #[inline]
-    pub fn get(&self, key: &str) -> Option<&PropertyValue> {
-        let interned_key = GLOBAL_INTERNER.get_id(key)?;
-        self.get_by_interned_key(&interned_key)
-    }
-
-    /// Get a property value by an already-interned key.
-    #[inline]
-    pub fn get_by_interned_key(&self, key: &PropertyKey) -> Option<&PropertyValue> {
-        self.inner.get(key)
-    }
-
-    /// Check if a property exists.
-    #[inline]
-    pub fn contains_key(&self, key: &str) -> bool {
-        let Some(interned_key) = GLOBAL_INTERNER.get_id(key) else {
-            return false;
-        };
-        self.contains_interned_key(&interned_key)
-    }
-
-    /// Check if a property exists by an already-interned key.
-    #[inline]
-    pub fn contains_interned_key(&self, key: &PropertyKey) -> bool {
-        self.inner.contains_key(key)
-    }
-
-    /// Get the number of properties.
-    #[inline]
-    pub fn len(&self) -> usize {
-        self.inner.len()
-    }
-
-    /// Check if the property map is empty.
-    #[inline]
-    pub fn is_empty(&self) -> bool {
-        self.inner.is_empty()
-    }
-
-    /// Iterate over all key-value pairs.
-    pub fn iter(&self) -> impl Iterator<Item = (&PropertyKey, &PropertyValue)> {
-        self.inner.iter()
-    }
-
-    /// Get all property keys.
-    pub fn keys(&self) -> impl Iterator<Item = &PropertyKey> {
-        self.inner.keys()
-    }
-
-    /// Get all property values.
-    pub fn values(&self) -> impl Iterator<Item = &PropertyValue> {
-        self.inner.values()
-    }
-
-    /// Create a builder for modifying this property map.
-    pub fn builder(self) -> PropertyMapBuilder {
-        PropertyMapBuilder::from_map(self)
-    }
-
-    /// Check if this property map contains any vector properties (dense or sparse).
-    #[inline]
-    pub fn contains_vector(&self) -> bool {
-        self.inner
-            .values()
-            .any(|v| matches!(v, PropertyValue::Vector(_) | PropertyValue::SparseVector(_)))
-    }
-
-    // ========================================================================
-    // Serialization Methods
-    // ========================================================================
-
-    /// Serialize this PropertyMap to bytes.
-    pub fn serialize(&self) -> Result<Vec<u8>> {
-        let mut buffer = Vec::with_capacity(self.cached_size);
-        self.serialize_into(&mut buffer)?;
-        Ok(buffer)
-    }
-
-    /// Serialize this PropertyMap into an existing buffer.
-    pub fn serialize_into(&self, buffer: &mut Vec<u8>) -> Result<()> {
-        buffer.reserve(self.cached_size);
-
-        buffer.extend_from_slice(&(self.inner.len() as u32).to_le_bytes());
-        for (key, value) in self.inner.iter() {
-            GLOBAL_INTERNER
-                .resolve_with(*key, |key_str| {
-                    let key_bytes = key_str.as_bytes();
-                    buffer.extend_from_slice(&(key_bytes.len() as u32).to_le_bytes());
-                    buffer.extend_from_slice(key_bytes);
-                })
-                .ok_or_else(|| {
-                    crate::utils::error::Error::Storage(StorageError::InconsistentState {
-                        reason: format!(
-                            "PropertyKey {} not found in interner - data corruption detected",
-                            key.as_u32()
-                        ),
-                    })
-                })?;
-
-            value.serialize_into(buffer)?;
-        }
-        Ok(())
-    }
-
-    /// Deserialize a PropertyMap from bytes.
-    pub fn deserialize(bytes: &[u8]) -> Result<(Self, usize)> {
-        if bytes.len() < 4 {
-            return Err(StorageError::CorruptedData(
-                "Buffer too short for PropertyMap count".to_string(),
-            )
-            .into());
-        }
-
-        let count = u32::from_le_bytes(bytes[0..4].try_into().unwrap()) as usize;
-
-        if count > MAX_PROPERTY_MAP_CAPACITY {
-            return Err(StorageError::CorruptedData(format!(
-                "PropertyMap count {} exceeds maximum allowed {}",
-                count, MAX_PROPERTY_MAP_CAPACITY
-            ))
-            .into());
-        }
-
-        let mut offset = 4;
-        let min_required_bytes = count.saturating_mul(5);
-        if bytes.len().saturating_sub(offset) < min_required_bytes {
-            return Err(StorageError::CorruptedData(format!(
-                "Insufficient buffer size for PropertyMap entries: need {} bytes, have {}",
-                min_required_bytes,
-                bytes.len().saturating_sub(offset)
-            ))
-            .into());
-        }
-
-        let mut map = HashMap::with_capacity(count);
-        let mut calculated_size: usize = 4;
-
-        for _ in 0..count {
-            if bytes.len() < offset + 4 {
-                return Err(StorageError::CorruptedData(
-                    "Buffer too short for property key length".to_string(),
-                )
-                .into());
-            }
-            let key_len =
-                u32::from_le_bytes(bytes[offset..offset + 4].try_into().unwrap()) as usize;
-            offset += 4;
-
-            if bytes.len() < offset + key_len {
-                return Err(StorageError::CorruptedData(
-                    "Buffer too short for property key data".to_string(),
-                )
-                .into());
-            }
-            let key_str = std::str::from_utf8(&bytes[offset..offset + key_len]).map_err(|e| {
-                StorageError::CorruptedData(format!("Invalid UTF-8 in property key: {}", e))
-            })?;
-            let key = GLOBAL_INTERNER.intern(key_str)?;
-            offset += key_len;
-
-            let (value, consumed) = PropertyValue::deserialize(&bytes[offset..])?;
-
-            let key_size = 4 + key_len;
-            calculated_size = calculated_size
-                .saturating_add(key_size)
-                .saturating_add(value.serialized_size()?);
-
-            offset += consumed;
-
-            if map.insert(key, value).is_some() {
-                return Err(StorageError::CorruptedData(format!(
-                    "Duplicate property key found during deserialization: '{}'. \
-                     This indicates corrupted data or invalid serialization format.",
-                    key_str
-                ))
-                .into());
-            }
-        }
-
-        if offset != calculated_size {
-            return Err(StorageError::CorruptedData(format!(
-                "PropertyMap deserialization size mismatch: consumed {} bytes but logical size is {}. \
-                 Data corruption suspected.",
-                offset, calculated_size
-            ))
-            .into());
-        }
-
-        Ok((
-            PropertyMap {
-                inner: Arc::new(map),
-                cached_size: calculated_size,
-            },
-            offset,
-        ))
-    }
-
-    /// Estimate the heap memory usage of this property map in bytes.
-    pub fn estimated_heap_size(&self) -> usize {
-        let mut size = self.inner.capacity()
-            * (std::mem::size_of::<PropertyKey>() + std::mem::size_of::<PropertyValue>() + 8);
-
-        for value in self.inner.values() {
-            size += value.estimated_heap_size();
-        }
-
-        size
-    }
-
-    /// Calculate the number of bytes required for serialization.
-    #[inline(always)]
-    pub fn serialized_size(&self) -> usize {
-        self.cached_size
-    }
-}
-
-impl Default for PropertyMap {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-impl FromIterator<(PropertyKey, PropertyValue)> for PropertyMap {
-    fn from_iter<I: IntoIterator<Item = (PropertyKey, PropertyValue)>>(iter: I) -> Self {
-        let mut map = HashMap::new();
-        let mut size: usize = 4; // Count field
-
-        for (key, value) in iter {
-            let key_len = GLOBAL_INTERNER
-                .resolve_with(key, |s| s.len())
-                .unwrap_or(256);
-            let key_size = 4 + key_len;
-            let val_size = value
-                .serialized_size()
-                .expect("Recursion depth limit exceeded in FromIterator");
-
-            size = size.saturating_add(key_size).saturating_add(val_size);
-
-            if let Some(old_val) = map.insert(key, value) {
-                size = size.saturating_sub(key_size).saturating_sub(
-                    old_val
-                        .serialized_size()
-                        .expect("Recursion depth limit exceeded in FromIterator"),
-                );
-            }
-        }
-
-        PropertyMap {
-            inner: Arc::new(map),
-            cached_size: size,
-        }
-    }
-}
-
-/// Builder for creating or modifying property maps with copy-on-write semantics.
-pub struct PropertyMapBuilder {
-    map: HashMap<PropertyKey, PropertyValue>,
-    current_size: usize,
-}
-
-impl PropertyMapBuilder {
-    /// Create a new builder with an empty map.
-    pub fn new() -> Self {
-        PropertyMapBuilder {
-            map: HashMap::new(),
-            current_size: 4,
-        }
-    }
-
-    /// Create a builder from an existing PropertyMap.
-    pub fn from_map(prop_map: PropertyMap) -> Self {
-        let current_size = prop_map.cached_size;
-        let map = Arc::try_unwrap(prop_map.inner).unwrap_or_else(|arc| (*arc).clone());
-        PropertyMapBuilder { map, current_size }
-    }
-
-    /// Insert a property.
-    pub fn insert<V: Into<PropertyValue>>(self, key: &str, value: V) -> Self {
-        self.try_insert(key, value)
-            .expect("Property insertion failed (recursion depth limit exceeded)")
-    }
-
-    /// Insert a property (fallible).
-    pub fn try_insert<V: Into<PropertyValue>>(mut self, key: &str, value: V) -> Result<Self> {
-        let Ok(interned_key) = GLOBAL_INTERNER.intern(key) else {
-            return Ok(self);
-        };
-        let val = value.into();
-        let val_size = val.serialized_size()?;
-
-        if let Some(old_val) = self.map.insert(interned_key, val) {
-            self.current_size = self
-                .current_size
-                .saturating_sub(old_val.serialized_size()?)
-                .saturating_add(val_size);
-        } else {
-            let key_size = 4 + key.len();
-            self.current_size = self
-                .current_size
-                .saturating_add(key_size)
-                .saturating_add(val_size);
-        }
-        Ok(self)
-    }
-
-    /// Insert a property with an already-interned key.
-    pub fn insert_by_key(self, key: PropertyKey, value: PropertyValue) -> Self {
-        self.try_insert_by_key(key, value)
-            .expect("Property insertion failed (recursion depth limit exceeded)")
-    }
-
-    /// Insert a property with an already-interned key (fallible).
-    pub fn try_insert_by_key(mut self, key: PropertyKey, value: PropertyValue) -> Result<Self> {
-        let val_size = value.serialized_size()?;
-
-        if let Some(old_val) = self.map.insert(key, value) {
-            self.current_size = self
-                .current_size
-                .saturating_sub(old_val.serialized_size()?)
-                .saturating_add(val_size);
-        } else {
-            let key_len = GLOBAL_INTERNER
-                .resolve_with(key, |s| s.len())
-                .unwrap_or_else(|| {
-                    debug_assert!(false, "PropertyKey {} missing from interner", key.as_u32());
-                    256
-                });
-            let key_size = 4 + key_len;
-            self.current_size = self
-                .current_size
-                .saturating_add(key_size)
-                .saturating_add(val_size);
-        }
-        Ok(self)
-    }
-
-    /// Insert a vector property.
-    pub fn insert_vector(self, key: &str, vector: &[f32]) -> Self {
-        self.insert(key, PropertyValue::vector(vector))
-    }
-
-    /// Insert a vector property (fallible).
-    pub fn try_insert_vector(self, key: &str, vector: &[f32]) -> Result<Self> {
-        self.try_insert(key, PropertyValue::vector(vector))
-    }
-
-    /// Remove a property.
-    pub fn remove(self, key: &str) -> Self {
-        self.try_remove(key).expect("Property removal failed")
-    }
-
-    /// Remove a property (fallible).
-    pub fn try_remove(self, key: &str) -> Result<Self> {
-        let Some(interned_key) = GLOBAL_INTERNER.get_id(key) else {
-            return Ok(self);
-        };
-        self.try_remove_by_key(&interned_key)
-    }
-
-    /// Remove a property by an already-interned key.
-    pub fn remove_by_key(self, key: &PropertyKey) -> Self {
-        self.try_remove_by_key(key)
-            .expect("Property removal failed")
-    }
-
-    /// Remove a property by an already-interned key (fallible).
-    pub fn try_remove_by_key(mut self, key: &PropertyKey) -> Result<Self> {
-        let old_val = self.map.remove(key);
-        if let Some(old_val) = old_val {
-            let key_len = GLOBAL_INTERNER
-                .resolve_with(*key, |s| s.len())
-                .unwrap_or(256);
-            let key_size = 4 + key_len;
-            self.current_size = self
-                .current_size
-                .saturating_sub(key_size)
-                .saturating_sub(old_val.serialized_size()?);
-        }
-        Ok(self)
-    }
-
-    /// Build the final PropertyMap.
-    pub fn build(self) -> PropertyMap {
-        PropertyMap {
-            inner: Arc::new(self.map),
-            cached_size: self.current_size,
-        }
-    }
-}
-
-impl Default for PropertyMapBuilder {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-/// Macro for creating property maps with a convenient syntax.
-#[macro_export]
-macro_rules! properties {
-    ($($key:expr => $value:expr),* $(,)?) => {
-        {
-            let mut builder = $crate::core::property::PropertyMapBuilder::new();
-            $(
-                builder = builder.insert($key, $value);
-            )*
-            builder.build()
-        }
-    };
-}
