--- src/core/property/serialization.rs
@@ -1,602 +1,0 @@
-use super::types::*;
-use crate::core::vector::SparseVec;
-use crate::utils::error::{Result, StorageError};
-use std::sync::Arc;
-
-/// Serialization logic for PropertyValue.
-pub struct PropertyValueSerialization;
-
-impl PropertyValueSerialization {
-    /// Serialize a PropertyValue to bytes.
-    pub fn serialize(value: &PropertyValue) -> Result<Vec<u8>> {
-        let mut buffer = Vec::with_capacity(Self::serialized_size(value).map_err(|_| {
-            StorageError::CorruptedData(
-                "Recursion depth limit exceeded in serialized_size".to_string(),
-            )
-        })?);
-        Self::serialize_into(value, &mut buffer)?;
-        Ok(buffer)
-    }
-
-    /// Serialize a PropertyValue into an existing buffer.
-    pub fn serialize_into(value: &PropertyValue, buffer: &mut Vec<u8>) -> Result<()> {
-        Self::serialize_recursive(value, buffer, 0)
-    }
-
-    /// Internal recursive serialization helper.
-    pub fn serialize_recursive(
-        value: &PropertyValue,
-        buffer: &mut Vec<u8>,
-        depth: usize,
-    ) -> Result<()> {
-        if depth > MAX_RECURSION_DEPTH {
-            return Err(StorageError::CorruptedData(format!(
-                "Property value recursion depth limit exceeded (max {})",
-                MAX_RECURSION_DEPTH
-            ))
-            .into());
-        }
-
-        match value {
-            PropertyValue::Null => {
-                buffer.push(TAG_NULL);
-                Ok(())
-            }
-            PropertyValue::Bool(b) => {
-                buffer.push(TAG_BOOL);
-                buffer.push(if *b { 1 } else { 0 });
-                Ok(())
-            }
-            PropertyValue::Int(i) => {
-                buffer.push(TAG_INT);
-                buffer.extend_from_slice(&i.to_le_bytes());
-                Ok(())
-            }
-            PropertyValue::Float(f) => {
-                buffer.push(TAG_FLOAT);
-                buffer.extend_from_slice(&f.to_le_bytes());
-                Ok(())
-            }
-            PropertyValue::String(s) => {
-                buffer.push(TAG_STRING);
-                let bytes = s.as_bytes();
-                buffer.extend_from_slice(&(bytes.len() as u32).to_le_bytes());
-                buffer.extend_from_slice(bytes);
-                Ok(())
-            }
-            PropertyValue::Bytes(b) => {
-                buffer.push(TAG_BYTES);
-                buffer.extend_from_slice(&(b.len() as u32).to_le_bytes());
-                buffer.extend_from_slice(b);
-                Ok(())
-            }
-            PropertyValue::Array(arr) => {
-                buffer.push(TAG_ARRAY);
-                buffer.extend_from_slice(&(arr.len() as u32).to_le_bytes());
-                for item in arr.iter() {
-                    // Call back into the type which delegates back to us
-                    // This maintains the encapsulation of PropertyValue having the method
-                    item.serialize_recursive(buffer, depth + 1)?;
-                }
-                Ok(())
-            }
-            PropertyValue::Vector(v) => {
-                serialize_vector_into(v, buffer);
-                Ok(())
-            }
-            PropertyValue::SparseVector(sv) => {
-                serialize_sparse_vector_into(sv, buffer);
-                Ok(())
-            }
-        }
-    }
-
-    /// Deserialize a PropertyValue from bytes.
-    pub fn deserialize(bytes: &[u8]) -> Result<(PropertyValue, usize)> {
-        Self::deserialize_recursive(bytes, 0)
-    }
-
-    /// Internal recursive deserialization helper.
-    pub fn deserialize_recursive(bytes: &[u8], depth: usize) -> Result<(PropertyValue, usize)> {
-        // Prevent recursion-based stack overflow DoS
-        // Depth 0 = top level, depth 100 = maximum nesting level
-        if depth > MAX_RECURSION_DEPTH {
-            return Err(StorageError::CorruptedData(format!(
-                "Property value recursion depth limit exceeded (max {})",
-                MAX_RECURSION_DEPTH
-            ))
-            .into());
-        }
-
-        if bytes.is_empty() {
-            return Err(StorageError::CorruptedData(
-                "Empty buffer when deserializing PropertyValue".to_string(),
-            )
-            .into());
-        }
-
-        let tag = bytes[0];
-        let mut offset = 1;
-
-        match tag {
-            TAG_NULL => Ok((PropertyValue::Null, offset)),
-
-            TAG_BOOL => {
-                if bytes.len() < 2 {
-                    return Err(StorageError::CorruptedData(
-                        "Buffer too short for Bool value".to_string(),
-                    )
-                    .into());
-                }
-                let value = bytes[1] != 0;
-                Ok((PropertyValue::Bool(value), 2))
-            }
-
-            TAG_INT => {
-                if bytes.len() < 9 {
-                    return Err(StorageError::CorruptedData(
-                        "Buffer too short for Int value".to_string(),
-                    )
-                    .into());
-                }
-                // SAFETY: Length check above guarantees slice has 8 bytes
-                let value = i64::from_le_bytes(bytes[1..9].try_into().unwrap());
-                Ok((PropertyValue::Int(value), 9))
-            }
-
-            TAG_FLOAT => {
-                if bytes.len() < 9 {
-                    return Err(StorageError::CorruptedData(
-                        "Buffer too short for Float value".to_string(),
-                    )
-                    .into());
-                }
-                // SAFETY: Length check above guarantees slice has 8 bytes
-                let value = f64::from_le_bytes(bytes[1..9].try_into().unwrap());
-                Ok((PropertyValue::Float(value), 9))
-            }
-
-            TAG_STRING => {
-                if bytes.len() < 5 {
-                    return Err(StorageError::CorruptedData(
-                        "Buffer too short for String length".to_string(),
-                    )
-                    .into());
-                }
-                let len = u32::from_le_bytes(bytes[1..5].try_into().unwrap()) as usize;
-                offset = 5;
-
-                if bytes.len() < offset + len {
-                    return Err(StorageError::CorruptedData(format!(
-                        "Buffer too short for String data: need {} bytes, have {}",
-                        offset + len,
-                        bytes.len()
-                    ))
-                    .into());
-                }
-
-                let string_data = &bytes[offset..offset + len];
-                let s = std::str::from_utf8(string_data).map_err(|e| {
-                    StorageError::CorruptedData(format!("Invalid UTF-8 in String: {}", e))
-                })?;
-                Ok((PropertyValue::String(Arc::from(s)), offset + len))
-            }
-
-            TAG_BYTES => {
-                if bytes.len() < 5 {
-                    return Err(StorageError::CorruptedData(
-                        "Buffer too short for Bytes length".to_string(),
-                    )
-                    .into());
-                }
-                let len = u32::from_le_bytes(bytes[1..5].try_into().unwrap()) as usize;
-                offset = 5;
-
-                if bytes.len() < offset + len {
-                    return Err(StorageError::CorruptedData(format!(
-                        "Buffer too short for Bytes data: need {} bytes, have {}",
-                        offset + len,
-                        bytes.len()
-                    ))
-                    .into());
-                }
-
-                let byte_data = &bytes[offset..offset + len];
-                Ok((PropertyValue::Bytes(Arc::from(byte_data)), offset + len))
-            }
-
-            TAG_ARRAY => {
-                if bytes.len() < 5 {
-                    return Err(StorageError::CorruptedData(
-                        "Buffer too short for Array count".to_string(),
-                    )
-                    .into());
-                }
-                let count = u32::from_le_bytes(bytes[1..5].try_into().unwrap()) as usize;
-                offset = 5;
-
-                // Prevent DoS via memory exhaustion from malicious input
-                if count > MAX_ARRAY_ELEMENTS {
-                    return Err(StorageError::CorruptedData(format!(
-                        "Array count {} exceeds maximum allowed {}",
-                        count, MAX_ARRAY_ELEMENTS
-                    ))
-                    .into());
-                }
-
-                // Prevent DoS via pre-allocation amplification
-                if bytes.len().saturating_sub(offset) < count {
-                    return Err(StorageError::CorruptedData(format!(
-                        "Insufficient buffer size for Array elements: need {} bytes, have {}",
-                        count,
-                        bytes.len().saturating_sub(offset)
-                    ))
-                    .into());
-                }
-
-                let mut items = Vec::with_capacity(count);
-                for _ in 0..count {
-                    if offset >= bytes.len() {
-                        return Err(StorageError::CorruptedData(
-                            "Buffer exhausted while reading Array elements".to_string(),
-                        )
-                        .into());
-                    }
-                    // Recursive call with depth increment
-                    let (item, consumed) =
-                        PropertyValue::deserialize_recursive(&bytes[offset..], depth + 1)?;
-                    items.push(item);
-                    offset += consumed;
-                }
-                Ok((PropertyValue::Array(Arc::new(items)), offset))
-            }
-
-            TAG_VECTOR => {
-                let (vector, consumed) = deserialize_vector(bytes)?;
-                Ok((PropertyValue::Vector(vector), consumed))
-            }
-
-            TAG_SPARSE_VECTOR => {
-                let (sparse_vector, consumed) = deserialize_sparse_vector(bytes)?;
-                Ok((PropertyValue::SparseVector(sparse_vector), consumed))
-            }
-
-            _ => Err(StorageError::CorruptedData(format!(
-                "Unknown PropertyValue type tag: {}",
-                tag
-            ))
-            .into()),
-        }
-    }
-
-    /// Estimate heap size of a PropertyValue.
-    pub fn estimated_heap_size(value: &PropertyValue) -> usize {
-        Self::estimated_heap_size_recursive(value, 0).unwrap_or(10 * 1024 * 1024)
-    }
-
-    /// Internal recursive heap size estimation helper.
-    pub fn estimated_heap_size_recursive(value: &PropertyValue, depth: usize) -> Result<usize> {
-        if depth > MAX_RECURSION_DEPTH {
-            return Err(StorageError::CorruptedData(format!(
-                "Property value recursion depth limit exceeded (max {})",
-                MAX_RECURSION_DEPTH
-            ))
-            .into());
-        }
-
-        match value {
-            PropertyValue::Null
-            | PropertyValue::Bool(_)
-            | PropertyValue::Int(_)
-            | PropertyValue::Float(_) => Ok(0),
-            PropertyValue::String(s) => Ok(s.len()),
-            PropertyValue::Bytes(b) => Ok(b.len()),
-            PropertyValue::Array(arr) => {
-                // Vec capacity overhead + recursive element sizes
-                let mut size = arr.capacity() * std::mem::size_of::<PropertyValue>();
-                for item in arr.iter() {
-                    size += item.estimated_heap_size_recursive(depth + 1)?;
-                }
-                Ok(size)
-            }
-            PropertyValue::Vector(v) => Ok(v.len() * std::mem::size_of::<f32>()),
-            PropertyValue::SparseVector(sv) => {
-                // Indices + values + SparseVec struct overhead
-                Ok(
-                    sv.nnz() * (std::mem::size_of::<u32>() + std::mem::size_of::<f32>())
-                        + std::mem::size_of::<usize>(),
-                )
-            }
-        }
-    }
-
-    /// Calculate serialized size of a PropertyValue.
-    pub fn serialized_size(value: &PropertyValue) -> Result<usize> {
-        Self::serialized_size_recursive(value, 0)
-    }
-
-    /// Internal recursive serialized size calculation helper.
-    pub fn serialized_size_recursive(value: &PropertyValue, depth: usize) -> Result<usize> {
-        if depth > MAX_RECURSION_DEPTH {
-            return Err(StorageError::CorruptedData(format!(
-                "Property value recursion depth limit exceeded (max {})",
-                MAX_RECURSION_DEPTH
-            ))
-            .into());
-        }
-
-        match value {
-            PropertyValue::Null => Ok(1),
-            PropertyValue::Bool(_) => Ok(2),
-            PropertyValue::Int(_) => Ok(9),
-            PropertyValue::Float(_) => Ok(9),
-            PropertyValue::String(s) => Ok(1 + 4 + s.len()),
-            PropertyValue::Bytes(b) => Ok(1 + 4 + b.len()),
-            PropertyValue::Array(arr) => {
-                let mut elements_size = 0;
-                for v in arr.iter() {
-                    elements_size += v.serialized_size_recursive(depth + 1)?;
-                }
-                Ok(1 + 4 + elements_size)
-            }
-            PropertyValue::Vector(v) => Ok(1 + 4 + (v.len() * 4)),
-            PropertyValue::SparseVector(sv) => Ok(1 + 4 + 4 + (sv.nnz() * 8)),
-        }
-    }
-}
-
-// ============================================================================
-// Vector Serialization Functions
-// ============================================================================
-
-/// Serialize a vector (dense f32 array) to bytes.
-pub fn serialize_vector(v: &[f32]) -> Vec<u8> {
-    let mut buffer = Vec::with_capacity(1 + 4 + v.len() * 4);
-    serialize_vector_into(v, &mut buffer);
-    buffer
-}
-
-/// Serialize a vector into an existing buffer.
-pub fn serialize_vector_into(v: &[f32], buffer: &mut Vec<u8>) {
-    if v.len() > MAX_VECTOR_DIMENSIONS {
-        panic!(
-            "Vector dimension {} exceeds maximum allowed {}",
-            v.len(),
-            MAX_VECTOR_DIMENSIONS
-        );
-    }
-
-    let required_size = 1 + 4 + std::mem::size_of_val(v);
-    buffer.reserve(required_size);
-
-    buffer.push(TAG_VECTOR);
-    buffer.extend_from_slice(&(v.len() as u32).to_le_bytes());
-
-    #[cfg(target_endian = "little")]
-    {
-        let byte_slice = unsafe {
-            std::slice::from_raw_parts(v.as_ptr() as *const u8, std::mem::size_of_val(v))
-        };
-        buffer.extend_from_slice(byte_slice);
-    }
-
-    #[cfg(not(target_endian = "little"))]
-    {
-        for &value in v {
-            buffer.extend_from_slice(&value.to_le_bytes());
-        }
-    }
-}
-
-/// Deserialize a vector from bytes.
-pub fn deserialize_vector(bytes: &[u8]) -> Result<(Arc<[f32]>, usize)> {
-    if bytes.len() < 5 {
-        return Err(
-            StorageError::CorruptedData("Buffer too short for vector header".to_string()).into(),
-        );
-    }
-
-    let tag = bytes[0];
-    if tag != TAG_VECTOR {
-        return Err(StorageError::CorruptedData(format!(
-            "Expected vector type tag {}, got {}",
-            TAG_VECTOR, tag
-        ))
-        .into());
-    }
-
-    let dimension = u32::from_le_bytes(bytes[1..5].try_into().unwrap()) as usize;
-
-    if dimension > MAX_VECTOR_DIMENSIONS {
-        return Err(StorageError::CorruptedData(format!(
-            "Vector dimension {} exceeds maximum allowed {}",
-            dimension, MAX_VECTOR_DIMENSIONS
-        ))
-        .into());
-    }
-
-    let data_start: usize = 5;
-    let data_len = dimension
-        .checked_mul(4)
-        .ok_or_else(|| StorageError::CorruptedData("Vector dimension overflow".to_string()))?;
-    let total_len = data_start
-        .checked_add(data_len)
-        .ok_or_else(|| StorageError::CorruptedData("Vector size overflow".to_string()))?;
-
-    if bytes.len() < total_len {
-        return Err(StorageError::CorruptedData(format!(
-            "Buffer too short for vector data: need {} bytes, have {}",
-            total_len,
-            bytes.len()
-        ))
-        .into());
-    }
-
-    let data_slice = &bytes[data_start..total_len];
-
-    #[cfg(target_endian = "little")]
-    let values = {
-        let mut values = Vec::with_capacity(dimension);
-        if dimension > 0 {
-            unsafe {
-                let src_ptr = data_slice.as_ptr();
-                let dst_ptr = values.as_mut_ptr() as *mut u8;
-                std::ptr::copy_nonoverlapping(src_ptr, dst_ptr, data_slice.len());
-                values.set_len(dimension);
-            }
-        }
-        values
-    };
-
-    #[cfg(not(target_endian = "little"))]
-    let values = {
-        let mut values = Vec::with_capacity(dimension);
-        for chunk in data_slice.chunks_exact(4) {
-            values.push(f32::from_le_bytes(chunk.try_into().unwrap()));
-        }
-        values
-    };
-
-    Ok((Arc::from(values.into_boxed_slice()), total_len))
-}
-
-// ============================================================================
-// Sparse Vector Serialization Functions
-// ============================================================================
-
-/// Serialize a sparse vector to bytes.
-pub fn serialize_sparse_vector(sv: &SparseVec) -> Vec<u8> {
-    let mut buffer = Vec::with_capacity(1 + 4 + 4 + sv.nnz() * 8);
-    serialize_sparse_vector_into(sv, &mut buffer);
-    buffer
-}
-
-/// Serialize a sparse vector into an existing buffer.
-pub fn serialize_sparse_vector_into(sv: &SparseVec, buffer: &mut Vec<u8>) {
-    buffer.reserve(1 + 4 + 4 + sv.nnz() * 8);
-
-    buffer.push(TAG_SPARSE_VECTOR);
-    buffer.extend_from_slice(&(sv.dimension() as u32).to_le_bytes());
-    buffer.extend_from_slice(&(sv.nnz() as u32).to_le_bytes());
-
-    for &idx in sv.indices() {
-        buffer.extend_from_slice(&idx.to_le_bytes());
-    }
-
-    for &val in sv.values() {
-        buffer.extend_from_slice(&val.to_le_bytes());
-    }
-}
-
-/// Deserialize a sparse vector from bytes.
-pub fn deserialize_sparse_vector(bytes: &[u8]) -> Result<(Arc<SparseVec>, usize)> {
-    if bytes.len() < 9 {
-        return Err(StorageError::CorruptedData(
-            "Buffer too short for sparse vector header".to_string(),
-        )
-        .into());
-    }
-
-    let tag = bytes[0];
-    if tag != TAG_SPARSE_VECTOR {
-        return Err(StorageError::CorruptedData(format!(
-            "Expected sparse vector type tag {}, got {}",
-            TAG_SPARSE_VECTOR, tag
-        ))
-        .into());
-    }
-
-    let dimension = u32::from_le_bytes(bytes[1..5].try_into().unwrap());
-    let nnz = u32::from_le_bytes(bytes[5..9].try_into().unwrap()) as usize;
-
-    if nnz > dimension as usize {
-        return Err(StorageError::CorruptedData(format!(
-            "Sparse vector nnz {} exceeds dimension {}",
-            nnz, dimension
-        ))
-        .into());
-    }
-
-    if nnz > MAX_VECTOR_DIMENSIONS {
-        return Err(StorageError::CorruptedData(format!(
-            "Sparse vector nnz {} exceeds maximum allowed {}",
-            nnz, MAX_VECTOR_DIMENSIONS
-        ))
-        .into());
-    }
-
-    let data_start: usize = 9;
-    let indices_len = nnz
-        .checked_mul(4)
-        .ok_or_else(|| StorageError::CorruptedData("Sparse vector nnz overflow".to_string()))?;
-    let values_len = indices_len;
-    let total_len = data_start
-        .checked_add(indices_len)
-        .and_then(|x: usize| x.checked_add(values_len))
-        .ok_or_else(|| StorageError::CorruptedData("Sparse vector size overflow".to_string()))?;
-
-    if bytes.len() < total_len {
-        return Err(StorageError::CorruptedData(format!(
-            "Buffer too short for sparse vector data: need {} bytes, have {}",
-            total_len,
-            bytes.len()
-        ))
-        .into());
-    }
-
-    let indices_end = data_start + indices_len;
-    let indices_slice = &bytes[data_start..indices_end];
-
-    #[cfg(target_endian = "little")]
-    let indices = {
-        let mut indices = Vec::with_capacity(nnz);
-        if nnz > 0 {
-            unsafe {
-                let src_ptr = indices_slice.as_ptr();
-                let dst_ptr = indices.as_mut_ptr() as *mut u8;
-                std::ptr::copy_nonoverlapping(src_ptr, dst_ptr, indices_slice.len());
-                indices.set_len(nnz);
-            }
-        }
-        indices
-    };
-
-    #[cfg(not(target_endian = "little"))]
-    let indices = {
-        let mut indices = Vec::with_capacity(nnz);
-        for chunk in indices_slice.chunks_exact(4) {
-            indices.push(u32::from_le_bytes(chunk.try_into().unwrap()));
-        }
-        indices
-    };
-
-    let values_end = indices_end + values_len;
-    let values_slice = &bytes[indices_end..values_end];
-
-    #[cfg(target_endian = "little")]
-    let values = {
-        let mut values = Vec::with_capacity(nnz);
-        if nnz > 0 {
-            unsafe {
-                let src_ptr = values_slice.as_ptr();
-                let dst_ptr = values.as_mut_ptr() as *mut u8;
-                std::ptr::copy_nonoverlapping(src_ptr, dst_ptr, values_slice.len());
-                values.set_len(nnz);
-            }
-        }
-        values
-    };
-
-    #[cfg(not(target_endian = "little"))]
-    let values = {
-        let mut values = Vec::with_capacity(nnz);
-        for chunk in values_slice.chunks_exact(4) {
-            values.push(f32::from_le_bytes(chunk.try_into().unwrap()));
-        }
-        values
-    };
-
-    let sparse_vec = SparseVec::new(indices, values, dimension)?;
-
-    Ok((Arc::new(sparse_vec), total_len))
-}
