--- src/core/property/types.rs
@@ -1,388 +1,0 @@
-use std::fmt;
-use std::sync::Arc;
-
-use crate::core::interning::InternedString;
-use crate::core::vector::SparseVec;
-use crate::utils::error::Result;
-
-// Forward declaration to resolve circular dependency eventually
-// use crate::core::property::serialization;
-
-// ============================================================================
-// Serialization Type Tags
-// ============================================================================
-// Binary format uses little-endian byte order (consistent with WAL format).
-// Each PropertyValue variant has a unique 1-byte type tag.
-
-/// Type tag for Null value.
-pub const TAG_NULL: u8 = 0;
-/// Type tag for Bool value.
-pub const TAG_BOOL: u8 = 1;
-/// Type tag for Int (i64) value.
-pub const TAG_INT: u8 = 2;
-/// Type tag for Float (f64) value.
-pub const TAG_FLOAT: u8 = 3;
-/// Type tag for String value.
-pub const TAG_STRING: u8 = 4;
-/// Type tag for Bytes value.
-pub const TAG_BYTES: u8 = 5;
-/// Type tag for Array value.
-pub const TAG_ARRAY: u8 = 6;
-/// Type tag for Vector (dense f32 array) value.
-pub const TAG_VECTOR: u8 = 7;
-/// Type tag for SparseVector value.
-pub const TAG_SPARSE_VECTOR: u8 = 8;
-
-// ============================================================================
-// Serialization Limits
-// ============================================================================
-// These limits prevent DoS attacks via memory exhaustion from malicious input.
-
-/// Maximum number of elements allowed in a deserialized array.
-/// Set to 1 million elements - enough for any practical use case.
-pub const MAX_ARRAY_ELEMENTS: usize = 1_000_000;
-
-/// Maximum number of dimensions allowed in a deserialized vector.
-/// Set to 100,000 - far exceeds typical embedding sizes (384-4096 dimensions).
-pub const MAX_VECTOR_DIMENSIONS: usize = 100_000;
-
-/// Maximum capacity allowed for a deserialized property map.
-/// Set to 10,000 to prevent OOM DoS attacks via malicious count fields.
-pub const MAX_PROPERTY_MAP_CAPACITY: usize = 10_000;
-
-/// Maximum recursion depth for nested properties (e.g., arrays of arrays).
-/// Set to 100 to prevent stack overflow from malicious input.
-pub const MAX_RECURSION_DEPTH: usize = 100;
-
-/// Property key type.
-///
-/// Uses interned strings for memory efficiency and O(1) equality comparisons.
-/// Common keys like "name", "age", and "id" are deduplicated in memory.
-pub type PropertyKey = InternedString;
-
-/// A value that can be stored as a property.
-///
-/// All complex types (strings, bytes, arrays) use Arc for cheap cloning and sharing.
-#[derive(Debug, Clone, PartialEq)]
-pub enum PropertyValue {
-    /// Null/absent value.
-    Null,
-    /// Boolean value.
-    Bool(bool),
-    /// 64-bit signed integer.
-    Int(i64),
-    /// 64-bit floating point.
-    Float(f64),
-    /// UTF-8 string (reference counted).
-    String(Arc<str>),
-    /// Byte array (reference counted).
-    Bytes(Arc<[u8]>),
-    /// Array of values (reference counted).
-    Array(Arc<Vec<PropertyValue>>),
-    /// Dense vector for embeddings (reference counted).
-    /// Uses f32 for memory efficiency - standard for ML embeddings.
-    Vector(Arc<[f32]>),
-    /// Sparse vector for high-dimensional sparse embeddings (reference counted).
-    SparseVector(Arc<SparseVec>),
-}
-
-impl PropertyValue {
-    /// Create a string property value from a &str.
-    #[inline]
-    pub fn string<S: AsRef<str>>(s: S) -> Self {
-        PropertyValue::String(Arc::from(s.as_ref()))
-    }
-
-    /// Create a bytes property value from a slice.
-    #[inline]
-    pub fn bytes<B: AsRef<[u8]>>(b: B) -> Self {
-        PropertyValue::Bytes(Arc::from(b.as_ref()))
-    }
-
-    /// Create an array property value from a Vec.
-    #[inline]
-    pub fn array(values: Vec<PropertyValue>) -> Self {
-        PropertyValue::Array(Arc::new(values))
-    }
-
-    /// Create a vector property value from a slice.
-    #[inline]
-    pub fn vector<V: AsRef<[f32]>>(v: V) -> Self {
-        let slice = v.as_ref();
-        if slice.len() > MAX_VECTOR_DIMENSIONS {
-            panic!(
-                "Vector dimension {} exceeds maximum allowed {}",
-                slice.len(),
-                MAX_VECTOR_DIMENSIONS
-            );
-        }
-        PropertyValue::Vector(Arc::from(slice))
-    }
-
-    /// Returns true if this value is null.
-    #[inline]
-    pub const fn is_null(&self) -> bool {
-        matches!(self, PropertyValue::Null)
-    }
-
-    /// Try to get this value as a bool.
-    #[inline]
-    pub const fn as_bool(&self) -> Option<bool> {
-        match self {
-            PropertyValue::Bool(b) => Some(*b),
-            _ => None,
-        }
-    }
-
-    /// Try to get this value as an integer.
-    #[inline]
-    pub const fn as_int(&self) -> Option<i64> {
-        match self {
-            PropertyValue::Int(i) => Some(*i),
-            _ => None,
-        }
-    }
-
-    /// Try to get this value as a float.
-    #[inline]
-    pub const fn as_float(&self) -> Option<f64> {
-        match self {
-            PropertyValue::Float(f) => Some(*f),
-            _ => None,
-        }
-    }
-
-    /// Try to get this value as a string reference.
-    #[inline]
-    pub fn as_str(&self) -> Option<&str> {
-        match self {
-            PropertyValue::String(s) => Some(s.as_ref()),
-            _ => None,
-        }
-    }
-
-    /// Try to get this value as a byte slice.
-    #[inline]
-    pub fn as_bytes(&self) -> Option<&[u8]> {
-        match self {
-            PropertyValue::Bytes(b) => Some(b.as_ref()),
-            _ => None,
-        }
-    }
-
-    /// Try to get this value as an array.
-    #[inline]
-    pub fn as_array(&self) -> Option<&[PropertyValue]> {
-        match self {
-            PropertyValue::Array(a) => Some(a.as_ref()),
-            _ => None,
-        }
-    }
-
-    /// Try to get this value as a vector (dense embedding).
-    #[inline]
-    pub fn as_vector(&self) -> Option<&[f32]> {
-        match self {
-            PropertyValue::Vector(v) => Some(v.as_ref()),
-            _ => None,
-        }
-    }
-
-    /// Get the underlying Arc for a vector (dense embedding) without copying.
-    #[inline]
-    pub fn as_arc_vector(&self) -> Option<Arc<[f32]>> {
-        match self {
-            PropertyValue::Vector(v) => Some(Arc::clone(v)),
-            _ => None,
-        }
-    }
-
-    /// Create a sparse vector property value from a SparseVec.
-    #[inline]
-    pub fn sparse_vector(sparse: SparseVec) -> Self {
-        PropertyValue::SparseVector(Arc::new(sparse))
-    }
-
-    /// Try to get this value as a sparse vector.
-    #[inline]
-    pub fn as_sparse_vector(&self) -> Option<&SparseVec> {
-        match self {
-            PropertyValue::SparseVector(sv) => Some(sv.as_ref()),
-            _ => None,
-        }
-    }
-
-    /// Get the type name of this value.
-    pub const fn type_name(&self) -> &'static str {
-        match self {
-            PropertyValue::Null => "null",
-            PropertyValue::Bool(_) => "bool",
-            PropertyValue::Int(_) => "int",
-            PropertyValue::Float(_) => "float",
-            PropertyValue::String(_) => "string",
-            PropertyValue::Bytes(_) => "bytes",
-            PropertyValue::Array(_) => "array",
-            PropertyValue::Vector(_) => "vector",
-            PropertyValue::SparseVector(_) => "sparse_vector",
-        }
-    }
-}
-
-impl fmt::Display for PropertyValue {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            PropertyValue::Null => write!(f, "null"),
-            PropertyValue::Bool(b) => write!(f, "{}", b),
-            PropertyValue::Int(i) => write!(f, "{}", i),
-            PropertyValue::Float(fl) => write!(f, "{}", fl),
-            PropertyValue::String(s) => write!(f, "\"{}\"", s),
-            PropertyValue::Bytes(b) => write!(f, "<{} bytes>", b.len()),
-            PropertyValue::Array(a) => {
-                write!(f, "[")?;
-                for (i, v) in a.iter().enumerate() {
-                    if i > 0 {
-                        write!(f, ", ")?;
-                    }
-                    write!(f, "{}", v)?;
-                }
-                write!(f, "]")
-            }
-            PropertyValue::Vector(v) => write!(f, "<vector[{}]>", v.len()),
-            PropertyValue::SparseVector(sv) => {
-                write!(
-                    f,
-                    "<sparse_vector[dim={}, nnz={}]>",
-                    sv.dimension(),
-                    sv.nnz()
-                )
-            }
-        }
-    }
-}
-
-// Convenient From implementations
-impl From<bool> for PropertyValue {
-    fn from(b: bool) -> Self {
-        PropertyValue::Bool(b)
-    }
-}
-
-impl From<i64> for PropertyValue {
-    fn from(i: i64) -> Self {
-        PropertyValue::Int(i)
-    }
-}
-
-impl From<i32> for PropertyValue {
-    fn from(i: i32) -> Self {
-        PropertyValue::Int(i as i64)
-    }
-}
-
-impl From<f64> for PropertyValue {
-    fn from(f: f64) -> Self {
-        PropertyValue::Float(f)
-    }
-}
-
-impl From<String> for PropertyValue {
-    fn from(s: String) -> Self {
-        PropertyValue::String(Arc::from(s))
-    }
-}
-
-impl From<&str> for PropertyValue {
-    fn from(s: &str) -> Self {
-        PropertyValue::String(Arc::from(s))
-    }
-}
-
-impl From<Vec<u8>> for PropertyValue {
-    fn from(b: Vec<u8>) -> Self {
-        PropertyValue::Bytes(Arc::from(b))
-    }
-}
-
-impl From<&[u8]> for PropertyValue {
-    fn from(b: &[u8]) -> Self {
-        PropertyValue::Bytes(Arc::from(b))
-    }
-}
-
-impl From<Vec<PropertyValue>> for PropertyValue {
-    fn from(v: Vec<PropertyValue>) -> Self {
-        PropertyValue::Array(Arc::new(v))
-    }
-}
-
-impl From<Vec<f32>> for PropertyValue {
-    fn from(v: Vec<f32>) -> Self {
-        PropertyValue::Vector(v.into())
-    }
-}
-
-impl From<&[f32]> for PropertyValue {
-    fn from(v: &[f32]) -> Self {
-        PropertyValue::Vector(Arc::from(v))
-    }
-}
-
-impl From<SparseVec> for PropertyValue {
-    fn from(sv: SparseVec) -> Self {
-        PropertyValue::SparseVector(Arc::new(sv))
-    }
-}
-
-// Serialization methods
-impl PropertyValue {
-    /// Serialize this PropertyValue to bytes.
-    pub fn serialize(&self) -> Result<Vec<u8>> {
-        use crate::core::property::serialization::PropertyValueSerialization;
-        PropertyValueSerialization::serialize(self)
-    }
-
-    /// Serialize this PropertyValue into an existing buffer.
-    pub fn serialize_into(&self, buffer: &mut Vec<u8>) -> Result<()> {
-        use crate::core::property::serialization::PropertyValueSerialization;
-        PropertyValueSerialization::serialize_into(self, buffer)
-    }
-
-    /// Deserialize a PropertyValue from bytes.
-    pub fn deserialize(bytes: &[u8]) -> Result<(Self, usize)> {
-        use crate::core::property::serialization::PropertyValueSerialization;
-        PropertyValueSerialization::deserialize(bytes)
-    }
-
-    /// Estimate the heap memory usage of this property value in bytes.
-    pub fn estimated_heap_size(&self) -> usize {
-        use crate::core::property::serialization::PropertyValueSerialization;
-        PropertyValueSerialization::estimated_heap_size(self)
-    }
-
-    /// Calculate the number of bytes required for serialization.
-    pub fn serialized_size(&self) -> Result<usize> {
-        use crate::core::property::serialization::PropertyValueSerialization;
-        PropertyValueSerialization::serialized_size(self)
-    }
-
-    // Helper to access private recursive methods from serialization module
-    pub(crate) fn serialize_recursive(&self, buffer: &mut Vec<u8>, depth: usize) -> Result<()> {
-        use crate::core::property::serialization::PropertyValueSerialization;
-        PropertyValueSerialization::serialize_recursive(self, buffer, depth)
-    }
-
-    pub(crate) fn deserialize_recursive(bytes: &[u8], depth: usize) -> Result<(Self, usize)> {
-        use crate::core::property::serialization::PropertyValueSerialization;
-        PropertyValueSerialization::deserialize_recursive(bytes, depth)
-    }
-
-    pub(crate) fn estimated_heap_size_recursive(&self, depth: usize) -> Result<usize> {
-        use crate::core::property::serialization::PropertyValueSerialization;
-        PropertyValueSerialization::estimated_heap_size_recursive(self, depth)
-    }
-
-    pub(crate) fn serialized_size_recursive(&self, depth: usize) -> Result<usize> {
-        use crate::core::property::serialization::PropertyValueSerialization;
-        PropertyValueSerialization::serialized_size_recursive(self, depth)
-    }
-}
